Moved AppIcon tests into the AppIcon component folder
