+++ src/components/AppIcon/tests/users-app-icon.png
